+++ app/lib/main.dart
@@ -60,25 +60,14 @@ Future<bool> _init() async {
   ServiceManager.init();
 
   // Firebase
-  if (kIsWeb) {
-    var options = F.env == Environment.prod
-        ? prod.DefaultFirebaseOptions.currentPlatform
-        : dev.DefaultFirebaseOptions.currentPlatform;
-    await Firebase.initializeApp(options: options);
-  } else if (F.env == Environment.prod) {
-    await Firebase.initializeApp(
-        options: prod.DefaultFirebaseOptions.currentPlatform, name: 'prod');
-  } else {
-    await Firebase.initializeApp(
-        options: dev.DefaultFirebaseOptions.currentPlatform, name: 'dev');
-  }
+  var options = F.env == Environment.prod
+      ? prod.DefaultFirebaseOptions.currentPlatform
+      : dev.DefaultFirebaseOptions.currentPlatform;
+  await Firebase.initializeApp(options: options);
 
-  SafeInit.init(() async {
-    await IntercomManager().initIntercom();
-    await NotificationService.instance.initialize();
-  }, () {
-    print("Notification service is not available for web");
-  });
+  await IntercomManager().initIntercom();
+  await NotificationService.instance.initialize();
+
   await SharedPreferencesUtil.init();
   await MixpanelManager.init();
 
@@ -92,6 +81,7 @@ Future<bool> _init() async {
   await GrowthbookUtil.init();
   CalendarUtil.init();
   SafeInit.init(() {
+    // not support on web.
     ble.FlutterBluePlus.setLogLevel(ble.LogLevel.info, color: true);
   });
   return isAuth;
@@ -137,8 +127,7 @@ void main() async {
           );
         }
         FlutterError.onError = (FlutterErrorDetails details) {
-          Zone.current.handleUncaughtError(
-              details.exception, details.stack ?? StackTrace.empty);
+          Zone.current.handleUncaughtError(details.exception, details.stack ?? StackTrace.empty);
         };
         Instabug.setColorTheme(ColorTheme.dark);
         runApp(const MyApp());
@@ -156,21 +145,17 @@ class MyApp extends StatefulWidget {
   @override
   State<MyApp> createState() => _MyAppState();
 
-  static _MyAppState of(BuildContext context) =>
-      context.findAncestorStateOfType<_MyAppState>()!;
+  static _MyAppState of(BuildContext context) => context.findAncestorStateOfType<_MyAppState>()!;
 
   // The navigator key is necessary to navigate using static methods
-  static final GlobalKey<NavigatorState> navigatorKey =
-      GlobalKey<NavigatorState>();
+  static final GlobalKey<NavigatorState> navigatorKey = GlobalKey<NavigatorState>();
 }
 
 class _MyAppState extends State<MyApp> with WidgetsBindingObserver {
   @override
   void initState() {
-    SafeInit.init(() {
-      NotificationUtil.initializeIsolateReceivePort();
-      NotificationUtil.initializeNotificationsEventListeners();
-    });
+    NotificationUtil.initializeIsolateReceivePort();
+    NotificationUtil.initializeNotificationsEventListeners();
     WidgetsBinding.instance.addObserver(this);
     super.initState();
   }
@@ -201,40 +186,31 @@ class _MyAppState extends State<MyApp> with WidgetsBindingObserver {
             update: (BuildContext context, value, MessageProvider? previous) =>
                 (previous?..updateAppProvider(value)) ?? MessageProvider(),
           ),
-          ChangeNotifierProxyProvider2<ConversationProvider, MessageProvider,
-              CaptureProvider>(
+          ChangeNotifierProxyProvider2<ConversationProvider, MessageProvider, CaptureProvider>(
             create: (context) => CaptureProvider(),
-            update: (BuildContext context, conversation, message,
-                    CaptureProvider? previous) =>
-                (previous?..updateProviderInstances(conversation, message)) ??
-                CaptureProvider(),
+            update: (BuildContext context, conversation, message, CaptureProvider? previous) =>
+                (previous?..updateProviderInstances(conversation, message)) ?? CaptureProvider(),
           ),
           ChangeNotifierProxyProvider<CaptureProvider, DeviceProvider>(
             create: (context) => DeviceProvider(),
-            update: (BuildContext context, captureProvider,
-                    DeviceProvider? previous) =>
+            update: (BuildContext context, captureProvider, DeviceProvider? previous) =>
                 (previous?..setProviders(captureProvider)) ?? DeviceProvider(),
           ),
           ChangeNotifierProxyProvider<DeviceProvider, OnboardingProvider>(
             create: (context) => OnboardingProvider(),
-            update: (BuildContext context, value,
-                    OnboardingProvider? previous) =>
+            update: (BuildContext context, value, OnboardingProvider? previous) =>
                 (previous?..setDeviceProvider(value)) ?? OnboardingProvider(),
           ),
           ListenableProvider(create: (context) => HomeProvider()),
           ChangeNotifierProxyProvider<DeviceProvider, SpeechProfileProvider>(
             create: (context) => SpeechProfileProvider(),
-            update: (BuildContext context, device,
-                    SpeechProfileProvider? previous) =>
+            update: (BuildContext context, device, SpeechProfileProvider? previous) =>
                 (previous?..setProviders(device)) ?? SpeechProfileProvider(),
           ),
-          ChangeNotifierProxyProvider2<AppProvider, ConversationProvider,
-              ConversationDetailProvider>(
+          ChangeNotifierProxyProvider2<AppProvider, ConversationProvider, ConversationDetailProvider>(
             create: (context) => ConversationDetailProvider(),
-            update: (BuildContext context, app, conversation,
-                    ConversationDetailProvider? previous) =>
-                (previous?..setProviders(app, conversation)) ??
-                ConversationDetailProvider(),
+            update: (BuildContext context, app, conversation, ConversationDetailProvider? previous) =>
+                (previous?..setProviders(app, conversation)) ?? ConversationDetailProvider(),
           ),
           ChangeNotifierProvider(create: (context) => CalenderProvider()),
           ChangeNotifierProvider(create: (context) => DeveloperModeProvider()),
@@ -271,20 +247,13 @@ class _MyAppState extends State<MyApp> with WidgetsBindingObserver {
                   ),
                   snackBarTheme: SnackBarThemeData(
                     backgroundColor: Colors.grey.shade900,
-                    contentTextStyle: const TextStyle(
-                        fontSize: 16,
-                        color: Colors.white,
-                        fontWeight: FontWeight.w500),
+                    contentTextStyle: const TextStyle(fontSize: 16, color: Colors.white, fontWeight: FontWeight.w500),
                   ),
                   textTheme: TextTheme(
-                    titleLarge:
-                        const TextStyle(fontSize: 18, color: Colors.white),
-                    titleMedium:
-                        const TextStyle(fontSize: 16, color: Colors.white),
-                    bodyMedium:
-                        const TextStyle(fontSize: 14, color: Colors.white),
-                    labelMedium:
-                        TextStyle(fontSize: 12, color: Colors.grey.shade200),
+                    titleLarge: const TextStyle(fontSize: 18, color: Colors.white),
+                    titleMedium: const TextStyle(fontSize: 16, color: Colors.white),
+                    bodyMedium: const TextStyle(fontSize: 14, color: Colors.white),
+                    labelMedium: TextStyle(fontSize: 12, color: Colors.grey.shade200),
                   ),
                   textSelectionTheme: const TextSelectionThemeData(
                     cursorColor: Colors.white,
@@ -292,20 +261,17 @@ class _MyAppState extends State<MyApp> with WidgetsBindingObserver {
                     selectionHandleColor: Colors.white,
                   ),
                   cupertinoOverrideTheme: const CupertinoThemeData(
-                    primaryColor:
-                        Colors.white, // Controls the selection handles on iOS
+                    primaryColor: Colors.white, // Controls the selection handles on iOS
                   )),
               themeMode: ThemeMode.dark,
               builder: (context, child) {
                 FlutterError.onError = (FlutterErrorDetails details) {
                   WidgetsBinding.instance.addPostFrameCallback((_) {
-                    Logger.instance.talker
-                        .handle(details.exception, details.stack);
+                    Logger.instance.talker.handle(details.exception, details.stack);
                   });
                 };
                 ErrorWidget.builder = (errorDetails) {
-                  return CustomErrorWidget(
-                      errorMessage: errorDetails.exceptionAsString());
+                  return CustomErrorWidget(errorMessage: errorDetails.exceptionAsString());
                 };
                 return child!;
               },
@@ -353,19 +319,15 @@ class _DeciderWidgetState extends State<DeciderWidget> {
   void openAppLink(Uri uri) async {
     if (uri.pathSegments.first == 'apps') {
       if (mounted) {
-        var app =
-            await context.read<AppProvider>().getAppFromId(uri.pathSegments[1]);
+        var app = await context.read<AppProvider>().getAppFromId(uri.pathSegments[1]);
         if (app != null) {
-          MixpanelManager()
-              .track('App Opened From DeepLink', properties: {'appId': app.id});
+          MixpanelManager().track('App Opened From DeepLink', properties: {'appId': app.id});
           if (mounted) {
-            Navigator.of(context).push(MaterialPageRoute(
-                builder: (context) => AppDetailPage(app: app)));
+            Navigator.of(context).push(MaterialPageRoute(builder: (context) => AppDetailPage(app: app)));
           }
         } else {
           debugPrint('App not found: ${uri.pathSegments[1]}');
-          AppSnackbar.showSnackbarError(
-              'Oops! Looks like the app you are looking for is not available.');
+          AppSnackbar.showSnackbarError('Oops! Looks like the app you are looking for is not available.');
         }
       }
     } else {
@@ -377,18 +339,16 @@ class _DeciderWidgetState extends State<DeciderWidget> {
   void initState() {
     SafeInit.init(() => initDeepLinks());
     WidgetsBinding.instance.addPostFrameCallback((_) async {
-      if (context.read<ConnectivityProvider>().isConnected && !kIsWeb) {
+      if (context.read<ConnectivityProvider>().isConnected) {
         NotificationService.instance.saveNotificationToken();
       }
 
       if (context.read<AuthenticationProvider>().isSignedIn()) {
         context.read<HomeProvider>().setupHasSpeakerProfile();
         try {
-          SafeInit.init(() async {
-            await IntercomManager.instance.intercom.loginIdentifiedUser(
-              userId: SharedPreferencesUtil().uid,
-            );
-          });
+          await IntercomManager.instance.intercom.loginIdentifiedUser(
+            userId: SharedPreferencesUtil().uid,
+          );
         } catch (e) {
           debugPrint('Failed to login to Intercom: $e');
         }
@@ -397,10 +357,9 @@ class _DeciderWidgetState extends State<DeciderWidget> {
         context.read<AppProvider>().setAppsFromCache();
         context.read<MessageProvider>().refreshMessages();
       } else {
-        SafeInit.init(() async =>
-            await IntercomManager.instance.intercom.loginUnidentifiedUser());
+        await IntercomManager.instance.intercom.loginUnidentifiedUser();
       }
-      SafeInit.init(() => IntercomManager.instance.setUserAttributes());
+      IntercomManager.instance.setUserAttributes();
     });
     super.initState();
   }
