+++ app/lib/pages/home/firmware_update.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:friend_private/utils/runtime.dart';
 import 'firmware_update_dialog.dart';
 import 'package:friend_private/backend/schema/bt_device/bt_device.dart';
 import 'package:friend_private/pages/home/firmware_mixin.dart';
@@ -158,9 +157,7 @@ class _FirmwareUpdateState extends State<FirmwareUpdate> with FirmwareMixin {
                                 if (updateMessage == '0')
                                   TextButton(
                                     onPressed: () async {
-                                      SafeInit.init(() async =>
-                                          await IntercomManager.instance
-                                              .displayFirmwareUpdateArticle());
+                                      await IntercomManager.instance.displayFirmwareUpdateArticle();
                                     },
                                     style: TextButton.styleFrom(
                                       minimumSize: Size.zero,
