+++ app/lib/pages/speech_profile/page.dart
@@ -12,7 +12,6 @@ import 'package:friend_private/providers/speech_profile_provider.dart';
 import 'package:friend_private/services/services.dart';
 import 'package:friend_private/utils/analytics/intercom.dart';
 import 'package:friend_private/utils/other/temp.dart';
-import 'package:friend_private/utils/runtime.dart';
 import 'package:friend_private/widgets/device_widget.dart';
 import 'package:friend_private/widgets/dialog.dart';
 import 'package:gradient_borders/box_borders/gradient_box_border.dart';
@@ -344,10 +343,7 @@ class _SpeechProfilePageState extends State<SpeechProfilePage> with TickerProvid
                                               context,
                                               () => Navigator.pop(context),
                                               () async {
-                                                SafeInit.init(() async {
-                                                  await IntercomManager.instance
-                                                      .displayFirmwareUpdateArticle();
-                                                });
+                                                await IntercomManager.instance.displayFirmwareUpdateArticle();
                                               },
                                               'Device Update Required',
                                               'Your current device has an old firmware version (1.0.2). Please check our guide on how to update it.',
