+++ app/lib/utils/analytics/analytics_manager.dart
@@ -1,6 +1,5 @@
 import 'package:friend_private/utils/analytics/intercom.dart';
 import 'package:friend_private/utils/analytics/mixpanel.dart';
-import 'package:friend_private/utils/runtime.dart';
 
 class AnalyticsManager {
   static final AnalyticsManager _instance = AnalyticsManager._internal();
@@ -13,15 +12,11 @@ class AnalyticsManager {
 
   void setUserAttributes() {
     MixpanelManager().setPeopleValues();
-    SafeInit.init(() {
-      IntercomManager.instance.setUserAttributes();
-    });
+    IntercomManager.instance.setUserAttributes();
   }
 
   void setUserAttribute(String key, dynamic value) {
     MixpanelManager().setUserProperty(key, value);
-    SafeInit.init(() {
-      IntercomManager.instance.updateCustomAttributes({key: value});
-    });
+    IntercomManager.instance.updateCustomAttributes({key: value});
   }
 }
