+++ app/web/index.html
@@ -140,31 +140,11 @@
   </script>
 </head>
 <body>
-  
-  
-  
-  
-  
-  <script src="https://www.gstatic.com/firebasejs/8.10.0/firebase-app.js"></script>
-  <script src="https://www.gstatic.com/firebasejs/8.10.0/firebase-auth.js"></script>
-  <script src="https://www.gstatic.com/firebasejs/8.10.0/firebase-messaging.js"></script>
 
 
-  <script>
-    var firebaseConfig = {
-      apiKey: "api_key",
-      authDomain: "auth.firebaseapp.com",
-      projectId: "project_id",
-      storageBucket: "storage..firebasestorage.app",
-      messagingSenderId: "sender_id",
-      appId: "api_id",
-      measurementId: "measurement_id",
-    };
 
-    // Initialize Firebase
-    firebase.initializeApp(firebaseConfig);
-  </script>
 
+  
   <script>
     window.addEventListener('load', function (ev) {
       // Download main.dart.js
